Fix rpt2 cache cleanup

+++ .travis-build.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 
 set -ex
-rm -r ./.rpt2_cache/
+[[ -e ./.rpt2_cache/ ]] && rm -r ./.rpt2_cache/
 node bundle.js
